+++ box-tool/src/main/java/smashdudes/boxtool/logic/SelectionType.java
@@ -1,0 +1,8 @@
+package smashdudes.boxtool.logic;
+
+public enum SelectionType
+{
+    None,
+    Move,
+    Scale,
+}
