+++ box-tool/src/main/java/smashdudes/boxtool/logic/commands/AddAnimationCommand.java
@@ -1,0 +1,28 @@
+package smashdudes.boxtool.logic.commands;
+
+import com.badlogic.gdx.utils.Array;
+import smashdudes.content.DTO;
+
+public class AddAnimationCommand extends Command
+{
+    private Array<DTO.Animation> animations;
+    private DTO.Animation anim;
+
+    public AddAnimationCommand(Array<DTO.Animation> animations, DTO.Animation anim)
+    {
+        this.animations = animations;
+        this.anim = anim;
+    }
+
+    @Override
+    protected void execute()
+    {
+        animations.add(anim);
+    }
+
+    @Override
+    protected void undo()
+    {
+        animations.removeValue(anim, true);
+    }
+}
