+++ box-tool/src/main/java/smashdudes/boxtool/logic/commands/Command.java
@@ -1,0 +1,8 @@
+package smashdudes.boxtool.logic.commands;
+
+public abstract class Command
+{
+    protected abstract void execute();
+
+    protected abstract void undo();
+}
