--- box-tool/src/main/java/smashdudes/core/boxtool/logic/SelectionType.java
@@ -1,8 +1,0 @@
-package smashdudes.core.boxtool.logic;
-
-public enum SelectionType
-{
-    None,
-    Move,
-    Scale,
-}
