--- box-tool/src/main/java/smashdudes/core/boxtool/logic/commands/ArraySwapCommand.java
@@ -1,30 +1,0 @@
-package smashdudes.core.boxtool.logic.commands;
-
-import com.badlogic.gdx.utils.Array;
-
-public class ArraySwapCommand extends Command
-{
-
-    private final Array list;
-    private final int idx0;
-    private final int idx1;
-
-    public ArraySwapCommand(Array list, int idx0, int idx1)
-    {
-        this.list = list;
-        this.idx0 = idx0;
-        this.idx1 = idx1;
-    }
-
-    @Override
-    protected void execute()
-    {
-        list.swap(idx0, idx1);
-    }
-
-    @Override
-    protected void undo()
-    {
-        list.swap(idx0, idx1);
-    }
-}
