--- box-tool/src/main/java/smashdudes/core/boxtool/logic/commands/Command.java
@@ -1,8 +1,0 @@
-package smashdudes.core.boxtool.logic.commands;
-
-public abstract class Command
-{
-    protected abstract void execute();
-
-    protected abstract void undo();
-}
